+++ R/defunct.R
@@ -1,0 +1,19 @@
+#' @name mschart-defunct
+#' @title Defunct Functions in Package mschart
+NULL
+
+#' @export
+#' @rdname mschart-defunct
+#' @details `ph_with_chart()` is replaced by `ph_with.ms_chart`.
+#' @param ... unused arguments
+ph_with_chart <- function(...) {
+  .Defunct("ph_with.ms_chart")
+}
+
+#' @export
+#' @rdname mschart-defunct
+#' @details `ph_with_chart_at()` is replaced by `ph_with.ms_chart`.
+ph_with_chart_at <- function(...) {
+  .Defunct("ph_with.ms_chart")
+}
+